+++ util/make_ramps.py
@@ -4,7 +4,7 @@ We use a pickled protocol=2, which is compat binary.
 """
 
 from pandas import read_sql
-from pyiem.util import get_dbconn
+from pyiem.util import get_dbconnstr
 
 PATH = "../src/pyiem/data/ramps/"
 # Be annoying
@@ -13,14 +13,12 @@ print("Be sure to run this against Mesonet database and not laptop!")
 
 def do(ramp):
     """states."""
-    pgconn = get_dbconn("mesosite")
-
     df = read_sql(
         "SELECT l.coloridx, l.value, l.r, l.g, l.b from iemrasters_lookup l "
         "JOIN iemrasters r ON (l.iemraster_id = r.id) WHERE r.name = %s and "
         "value is not null "
         "ORDER by coloridx ASC",
-        pgconn,
+        get_dbconnstr("mesosite"),
         params=(ramp,),
         index_col="coloridx",
     )