+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this library are documented in this file.
 - Add `pyiem.plot.geoplot.MapPlot` kwarg of `axes_position` to control where
 the main axes is placed on the figure.
 - Python versions 3.8 through 3.10 are now actively tested and supported.
+- Added `pyiem.util.get_dbconnstr` helper to return a database connection
+string details.
 
 ### Bug Fixes
 
@@ -29,6 +31,7 @@ which cross references VTEC codes to NWS WaWA map colors.
 - The NWS has an ongoing issue (2017-present) with disseminating truncated
 Tsuanmi products.  These truncated products are dups to non-truncated ones.
 Special code was added to abort processing of these when found.
+- A pandas 1.4.0 API problem was fixed.
 
 ## **1.9.0** (13 Dec 2021)